fix: Fix bugs.
Fix an emit diagnostic bug (because the location given to
the diagnostic was NULL).

Plus fix another bug in the return statement. The compiler does not allow
add a `return;` at the end of the function when the return data type
of the function was set to unknown. The compiler treats this as an error.