refactor(bin/lily): Call LLVMShutdown.
So far I haven't found a solution to prevent LLVM from loading and
allocating memory, even if it's not used in this program. So, for
the moment, we're calling LLVMShutdown for this reason. In the
end, I think the best solution to solve this problem and improve
program performance would be to stop using LLVM.