feat(cc/ci): Scan GNU attribute.
For the moment, we've only scanned the GNU attributes, but we're not
going to do anything with it, as it's just a matter of supporting the
syntax. So, in the future we'll probably have to parse the GNU
attributes and all that, but for now it's ok to do it this way.